--- src/views/ModelPlan/ReadOnly/_components/FilterViewBanner/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const FilterViewBanner = () => {
-  return (
-    <div className="position-sticky z-100 top-0 bg-primary-darker text-white padding-105">
-      FilterViewBanner
-    </div>
-  );
-};
-
-export default FilterViewBanner;
